WSS-17: Implementation of experiment statistics in renamed statistics module.
Created experiment statistics:
    * worker_throughput
    * system_throughput
    * makespan
