--- app/models/workers_scaling/statistics_interface.rb
@@ -1,32 +1,0 @@
-require 'experiment_extensions/simulation_run'
-
-class StatisticsInterface
-
-  def initialize(experiment_id)
-    @simulation_run = SimulationRun.for_experiment(experiment_id)
-  end
-
-  ##
-  # Returns SimulationRuns with given params
-  # params:
-  #   * opts
-  #   * cond
-  def for_simulation_runs(params = {})
-    query = params[:cond] || {}
-    opts= params[:opts] || {}
-    @simulation_run.where(query, opts).to_a
-  end
-
-  def worker_throughput(worker_id)
-
-  end
-
-  def system_throughput
-
-  end
-
-  def makespan
-
-  end
-
-end
